@@ -3,60 +3,150 @@
   "version": "1.6.2",
   "description": "Ramda validation utilities for the Rescape platform",
   "homepage": "https://www.rescapes.net/",
-  "main": "dist/index.js",
+  "main": "lib/index.js",
+  "module": "es/index.js",
   "files": [
-    "index.js",
-    "src",
-    "dist"
+    "dist",
+    "lib",
+    "es",
+    "src"
   ],
   "scripts": {
-    "test": "NODE_PATH='.:src' BABEL_ENV=test jest",
-    "test:coverage": "NODE_PATH='src' jest --coverage",
-    "lint": "eslint src example test",
-    "clean": "rm -fr dist/*",
-    "build": "npm run clean && npm run lint && npm run build-es5",
-    "build-es5": "mkdir -p dist && babel src --out-dir dist --ignore spec.js --plugins=transform-es2015-modules-commonjs --source-maps inline",
-    "build-publish-patch": "npm run build && npm run test && npm version patch && npm publish",
-    "build-publish-minor": "npm run build && npm run test && npm version minor && npm publish"
-  },
-  "jest": {
-    "modulePaths": [
-      "<rootDir>/source"
-    ],
-    "testPathIgnorePatterns": [
-      "__snapshots__"
-    ],
-    "bail": true
+    "clean": "rimraf lib dist es coverage",
+    "lint": "eslint src",
+    "test": "cross-env BABEL_ENV=commonjs jest",
+    "test:watch": "yarn test -- --watch",
+    "test:cov": "yarn test -- --coverage",
+    "build:commonjs": "cross-env NODE_ENV=production BABEL_ENV=commonjs  node_modules/.bin/babel src --out-dir lib",
+    "build:es": "cross-env  NODE_ENV=production BABEL_ENV=es node_modules/.bin/babel src --out-dir es",
+    "build:umd": "cross-env NODE_ENV=production BABEL_ENV=es NODE_ENV=development rollup -c -i src/index.js -o dist/index.js",
+    "build:umd:min": "cross-env NODE_ENV=production BABEL_ENV=es NODE_ENV=production rollup -c -i src/index.js -o dist/index.min.js",
+    "build": "yarn run build:commonjs && yarn run build:es && yarn run build:umd && yarn run build:umd:min",
+    "prepare": "yarn run clean && yarn run lint && yarn test && yarn run build",
+    "docs:clean": "rimraf _book",
+    "docs:prepare": "gitbook install",
+    "docs:build": "yarn run docs:prepare && gitbook build -g reactjs/redux && cp logo/apple-touch-icon.png _book/gitbook/images/apple-touch-icon-precomposed-152.png && cp logo/favicon.ico _book/gitbook/images",
+    "docs:watch": "yarn run docs:prepare && gitbook serve",
+    "docs:publish": "yarn run docs:clean && yarn run docs:build && cp CNAME _book && cd _book && git init && git commit --allow-empty -m 'update book' && git checkout -b gh-pages && touch .nojekyll && git add . && git commit -am 'update book' && git push git@github.com:reactjs/redux gh-pages --force",
+    "build-publish-patch": "yarn run build && yarn run publish-patch",
+    "build-publish-minor": "yarn run build && yarn run publish-minor",
+    "publish-patch": "yarn --new-version version patch && yarn publish",
+    "publish-minor": "yarn --new-version version minor && yarn publish"
   },
   "repository": "git@github.com:calocan/rescape-cycle.git",
   "author": "Andy Likuski",
   "license": "MIT",
   "dependencies": {
-    "@cycle/http": "^14.2.0",
-    "babel-preset-env": "^1.6.0",
+    "@cycle/http": "^14.9.0",
+    "babel-preset-env": "^1.7.0",
     "data.maybe": "^1.2.2",
-    "pretty-format": "^20.0.3",
-    "prop-types": "^15.5.10",
-    "ramda": "^0.24.1",
+    "prop-types": "^15.6.1",
+    "ramda": "^0.25.0",
     "ramda-fantasy": "^0.8.0",
     "ramda-fantasy-validation": "^0.2.0",
-    "redux": "^3.7.2",
+    "redux": "^4.0.0",
     "redux-cycles": "^0.4.1",
     "redux-thunk": "^2.2.0",
-    "rescape-ramda": "^1.2.8",
-    "rescape-validate": "^1.2.2",
-    "xstream": "^10.9.0"
+    "rescape-ramda": "^1.4.38",
+    "rescape-validate": "^1.2.9",
+    "xstream": "^11.3.0"
   },
   "devDependencies": {
-    "@cycle/time": "^0.8.0",
+    "@cycle/time": "^0.13.0",
     "babel-cli": "^6.26.0",
-    "babel-core": "6.25.0",
-    "babel-loader": "7.1.1",
-    "eslint": "^4.6.1",
-    "eslint-plugin-jest": "^21.0.0",
-    "jest": "^20.0.4",
-    "jsdom": "^11.1.0",
-    "redux-mock-store": "^1.3.0",
-    "webpack": "^3.5.4"
+    "babel-core": "6.26.3",
+    "babel-jest": "^22.4.4",
+    "babel-loader": "7.1.4",
+    "babel-plugin-external-helpers": "^6.22.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.7.0",
+    "cross-env": "^5.1.6",
+    "eslint": "^4.19.1",
+    "eslint-plugin-jest": "^21.15.2",
+    "jest": "^22.4.4",
+    "jsdom": "^11.11.0",
+    "redux-mock-store": "^1.5.1",
+    "rimraf": "^2.6.2",
+    "rollup": "^0.59.2",
+    "rollup-plugin-babel": "^3.0.4",
+    "rollup-plugin-node-resolve": "^3.3.0",
+    "rollup-plugin-replace": "^2.0.0",
+    "rollup-plugin-uglify": "^4.0.0"
+  },
+  "jest": {
+    "modulePaths": [
+      "<rootDir>/src"
+    ],
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx,mjs}"
+    ],
+    "testMatch": [
+      "<rootDir>/src/**/__tests__/**/*.{js,jsx,mjs}",
+      "<rootDir>/src/**/?(*.)(spec|test).{js,jsx,mjs}"
+    ],
+    "testEnvironment": "node",
+    "testURL": "http://localhost",
+    "transform": {
+      "^.+\\.(js|jsx|mjs)$": "<rootDir>/node_modules/babel-jest",
+      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
+      "^(?!.*\\.(js|jsx|mjs|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
+    },
+    "transformIgnorePatterns": [
+      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$"
+    ],
+    "moduleNameMapper": {
+      "^react-native$": "react-native-web"
+    },
+    "moduleFileExtensions": [
+      "web.js",
+      "mjs",
+      "js",
+      "json",
+      "web.jsx",
+      "jsx",
+      "node"
+    ]
+  },
+  "babel": {
+    "presets": [
+      [
+        "env",
+        {
+          "targets": {
+            "browsers": [
+              "ie >= 11"
+            ]
+          },
+          "exclude": [
+            "transform-async-to-generator"
+          ],
+          "modules": false,
+          "loose": true
+        }
+      ]
+    ],
+    "plugins": [
+      [
+        "transform-runtime",
+        {
+          "polyfill": false,
+          "regenerator": true
+        }
+      ]
+    ],
+    "env": {
+      "commonjs": {
+        "presets": [
+          [
+            "env",
+            {
+              "targets": {
+                "node": "current"
+              }
+            }
+          ]
+        ]
+      }
+    }
   }
 }
