@@ -8,17 +8,14 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-const R = require('ramda');
-const {mapDefault} = require('rescape-ramda');
-const {xs} = mapDefault('xs', require('xstream'));
-const {
-  VERBS: {FETCH, ADD, REMOVE, REPLACE, SELECT, DESELECT, MOVE, COPY},
-  PHASES: {REQUEST, SUCCESS, FAILURE},
-  PATCH_VERBS
-} = require('./helpers/actionHelpers');
-const {v} = require('rescape-validate');
-const PropTypes = require('prop-types');
-const prettyFormat = require('pretty-format');
+import R from 'ramda';
+import {mapDefault} from 'rescape-ramda';
+import xs from 'xstream'
+import { VERBS, PHASES, PATCH_VERBS } from './helpers/actionHelpers';
+import {v} from 'rescape-validate';
+import PropTypes from 'prop-types';
+const {FETCH, ADD, REMOVE, REPLACE, SELECT, DESELECT, MOVE, COPY} = VERBS
+const {REQUEST, SUCCESS, FAILURE} = PHASES
 
 /**
  * Convert incoming intents to cycle actions
